requirements.txt
# Dev environment
pip
autopep8

# App
langchain
langchain_community
chainlit
streamlit
transformers
sentence_transformers==2.2.2
tokenizers>=0.19,<0.20
ollama
datasets
faiss-cpu
huggingface_hub

hf_transfer
xmltodict
wikipedia

#agents dependencies
crewai
crewai_tools
ipython
python-dotenv
google-api-python-client